refactor(daemon,web): 技能模块去重统一——同步原语下沉 + 死代码清理
- 新增 skills/dirsync.ts：内容哈希比较 + tmp→rename 原子镜像，统一 library
  同步与 bundled 安装两条路径；skill-installer 删除私有 version 比较与
  copyDirSync（version 只作诊断，同步以哈希为准）
- skillmd 生成/解析从 daemon、web 双份镜像收敛到 @molio/contracts 单源
- knowledge.ts 抽 toVaultSkillEntry，消除 GET/PATCH 两处重复构造
- 清理死代码：manifestPath()、countSeededBuiltins()、removeSkillSyncDir、
  useSkills.prefill；i18n key claudeOnlyNote → runtimeNote
- 合并 test/tools/skill-installer.test.ts 进 test/core/，三份重复的
  ALL_BUNDLED/installAll 测试助手收敛到 test/helpers/install-all.ts
