+++ packages/contracts/src/index.ts
@@ -106,3 +106,7 @@ export type {
   VaultSkillListResponse,
   VaultSkillToggleRequest,
 } from './skill.js';
+
+// SKILL.md generate/parse primitives (runtime values, shared daemon ↔ web).
+export { generateSkillMd, stripFrontmatter, parseSkillMd } from './skillmd.js';
+export type { ParsedSkillMd } from './skillmd.js';
